Timeslide memory issue (#189)
* overhauling how waveforms get generated

* getting new timeslide implementation working

* updating ml4gw ref

* getting rid of priors and fixing test import
