+++ projects/sandbox/timeslide_injections/timeslide_injections/main.py
@@ -1,12 +1,8 @@
-import itertools
 import logging
-from concurrent.futures import FIRST_EXCEPTION, Future, wait
-from dataclasses import dataclass
+from concurrent.futures import FIRST_EXCEPTION, wait
 from pathlib import Path
-from typing import Callable, Dict, Iterable, List, Optional
+from typing import Callable, Iterable, Optional
 
-import bilby
-import gwdatafind
 import h5py
 import numpy as np
 import torch
@@ -16,186 +12,22 @@ from gwpy.segments import (
     SegmentList,
     SegmentListDict,
 )
-from gwpy.timeseries import TimeSeries, TimeSeriesDict
+from timeslide_injections.utils import (
+    Sampler,
+    WaveformGenerator,
+    check_segment,
+    download_data,
+    make_shifts,
+    submit_write,
+    waveform_iterator,
+)
 from typeo import scriptify
 
-from bbhnet.injection import generate_gw, inject_waveforms
-from bbhnet.io import h5
+from bbhnet.injection import inject_waveforms
 from bbhnet.io.timeslides import TimeSlide
 from bbhnet.logging import configure_logging
 from bbhnet.parallelize import AsyncExecutor
 from ml4gw.gw import compute_ifo_snr, compute_observed_strain, get_ifo_geometry
-
-
-@dataclass
-class Shift:
-    ifos: List[str]
-    shifts: Iterable[float]
-
-    def __post_init__(self):
-        self.shifts = [float(i) for i in self.shifts]
-        self._i = 0
-
-    def __iter__(self):
-        self._i = 0
-        return self
-
-    def __next__(self):
-        if self._i >= len(self.ifos):
-            raise StopIteration
-
-        ifo, shift = self.ifos[self._i], self.shifts[self._i]
-        self._i += 1
-        return ifo, shift
-
-    def __str__(self):
-        return "-".join([f"{i[0]}{j}" for i, j in zip(self.ifos, self.shifts)])
-
-
-def make_shifts(
-    ifos: Iterable[str], shifts: Iterable[float], n_slides: int
-) -> List[Shift]:
-    ranges = [range(n_slides) for i in shifts if i]
-    shift_objs = []
-    for rng in itertools.product(*ranges):
-        it = iter(rng)
-        shift = []
-        for i in shifts:
-            shift.append(0 if i == 0 else next(it) * i)
-        shift = Shift(ifos, shift)
-        shift_objs.append(shift)
-
-    return shift_objs
-
-
-def submit_write(
-    pool: AsyncExecutor, ts: TimeSlide, t: np.ndarray, **fields: np.ndarray
-) -> Future:
-    ts_type = ts.path.name
-    if ts_type == "background":
-        prefix = "raw"
-    else:
-        prefix = "inj"
-
-    future = pool.submit(
-        h5.write_timeseries,
-        ts.path,
-        prefix=prefix,
-        t=t,
-        **fields,
-    )
-
-    future.add_done_callback(
-        lambda f: logging.debug(f"Wrote background {ts_type} {f.result()}")
-    )
-    return future
-
-
-def download_data(
-    ifos: Iterable[str],
-    frame_type: str,
-    channel: str,
-    sample_rate: float,
-    start: float,
-    stop: float,
-) -> TimeSeriesDict:
-    data = TimeSeriesDict()
-    for ifo in ifos:
-        files = gwdatafind.find_urls(
-            site=ifo.strip("1"),
-            frametype=f"{ifo}_{frame_type}",
-            gpsstart=start,
-            gpsend=stop,
-            urltype="file",
-        )
-        data[ifo] = TimeSeries.read(
-            files, channel=f"{ifo}:{channel}", start=start, end=stop, nproc=4
-        )
-    return data.resample(sample_rate)
-
-
-def get_params(
-    priors: bilby.gw.prior.PriorDict,
-    n_slides: int,
-    signal_times: np.ndarray,
-    jitter: float,
-    waveform_duration: float,
-) -> List[Dict[str, List[float]]]:
-    n_signals = len(signal_times)
-    params = priors.sample(n_signals * n_slides)
-
-    shift_parameters = []
-    for i in range(n_slides):
-        slc = slice(i * n_signals, (i + 1) * n_signals)
-        param = {p: v[slc] for p, v in params.items()}
-
-        jit = np.random.uniform(-jitter, jitter, size=n_signals)
-        param["geocent_time"] = signal_times + jit + waveform_duration / 2
-
-        shift_parameters.append(param)
-    return shift_parameters
-
-
-def generate_waveforms(
-    parameters,
-    minimum_frequency,
-    reference_frequency,
-    sample_rate,
-    waveform_duration,
-    waveform_approximant,
-):
-    """
-    Cheap wrapper around waveform generation for async
-    execution so that we can know which parameters these
-    waveforms correspond to.
-    """
-
-    signals = generate_gw(
-        parameters,
-        minimum_frequency,
-        reference_frequency,
-        sample_rate,
-        waveform_duration,
-        waveform_approximant,
-    )
-    return signals, parameters
-
-
-def intify(x: float):
-    return int(x) if int(x) == x else x
-
-
-def check_segment(
-    shifts: List[Shift],
-    datadir: Path,
-    segment_start: float,
-    dur: float,
-    min_segment_length: Optional[float] = None,
-    force_generation: bool = False,
-):
-    # first check if we'll even have enough data for
-    # this segment to be worth working with
-    if min_segment_length is not None and dur < min_segment_length:
-        return None
-
-    segment_start = intify(segment_start)
-    dur = intify(dur)
-
-    # then check if _all_ data for this segment
-    # exists in each shift separately
-    fields, prefixes = ["background", "injection"], ["raw", "inj"]
-    segment_shifts = []
-    for shift in shifts:
-        for field, prefix in zip(fields, prefixes):
-            dirname = datadir / f"dt-{shift}" / field
-            fname = f"{prefix}_{segment_start}-{dur}.hdf5"
-            if not (dirname / fname).exists() or force_generation:
-                # we don't have data for this segment at this
-                # shift value, so we'll need to create it
-                segment_shifts.append(shift)
-                break
-
-    return segment_shifts
 
 
 @scriptify
@@ -294,9 +126,16 @@ def main(
     max_shift = max(shifts) * n_slides
     shifts = make_shifts(ifos, shifts, n_slides)
 
-    # grab some parameters we'll need for waveform generation
+    # grab some parameters we'll need for waveform injection
     stride = 1 / sample_rate
     priors = prior()
+    waveform_generator = WaveformGenerator(
+        minimum_frequency,
+        reference_frequency,
+        sample_rate,
+        waveform_duration,
+        waveform_approximant,
+    )
     tensors, vertices = get_ifo_geometry(*ifos)
 
     # set up some pools for doing our data IO/injection
@@ -323,32 +162,19 @@ def main(
                 continue
             num_shifts = len(segment_shifts)
 
-            # generate timestamps of trigger times for each waveform.
-            # Jitter will get added in `get_params`.
-            # take off min segment length from end of segment stop
-            # since we end up cutting some data off to make sure each timeslide
-            # has the same length of data
-            signal_start = segment_start + buffer_
-            signal_stop = segment_stop - buffer_ - max_shift
-            signal_times = np.arange(signal_start, signal_stop, spacing)
-
-            # sample dictionary of params for each timeslide
-            num_signals = num_shifts * len(signal_times)
-            logging.debug(f"Sampling {num_signals} waveform parameters")
-            parameters = get_params(
-                priors, num_shifts, signal_times, jitter, waveform_duration
+            # create an iterator which will generate raw
+            # waveforms in a separate process
+            sampler = Sampler(
+                priors,
+                segment_start,
+                segment_stop,
+                buffer_,
+                max_shift,
+                spacing,
+                jitter,
             )
-
-            # generate each timeslide's waveforms in parallel
-            logging.debug("Launching waveform generation in parallel")
-            waveform_it = pool.imap(
-                generate_waveforms,
-                parameters,
-                minimum_frequency=minimum_frequency,
-                reference_frequency=reference_frequency,
-                sample_rate=sample_rate,
-                waveform_duration=waveform_duration,
-                waveform_approximant=waveform_approximant,
+            waveform_it = waveform_iterator(
+                pool, sampler, waveform_generator, num_shifts
             )
 
             # begin the download of data in a separate thread
